Moving construction of the traverse routine to its own command, to decouple initialization and construction.
